chore: move getOrientationOfPoints into maths, skip orientation pass at default alignment
