+++ src/maths/index.ts
@@ -2,6 +2,7 @@
 export * from './matrix/groupD8';
 export * from './matrix/Matrix';
 export * from './misc/const';
+export * from './misc/getOrientationOfPoints';
 export * from './misc/pow2';
 export * from './misc/Size';
 export * from './misc/squaredDistanceToLineSegment';
